Extract message format contexts into helper
