+++ spec/functional/message_spec.rb
@@ -1,6 +1,5 @@
 require 'spec_helper'
-
-require 'avro/registered_schema_decoder'
+require 'format_contexts'
 
 shared_examples 'publishing messages' do |format|
   # We only stop the cluster after all examples in the context have run, so
@@ -119,41 +118,6 @@ shared_examples 'publishing messages' do |format|
   end
 end
 
-shared_context 'JSON format' do
-  def parse(*args); JSON.parse(*args); end
-  alias decode_key parse
-  alias decode_value parse
-
-  def fetch_int(object, name)
-    object.fetch(name).fetch('int')
-  end
-
-  def fetch_string(object, name)
-    object.fetch(name).fetch('string')
-  end
-end
-
-shared_context 'Avro format' do
-  let(:decoder) do
-    Avro::RegisteredSchemaDecoder.new(
-      TEST_CLUSTER.schema_registry_url,
-      logger: logger)
-  end
-
-  def decode_key(*args)
-    decoder.decode_key(*args)
-  end
-
-  def decode_value(*args)
-    decoder.decode_value(*args)
-  end
-
-  def fetch_entry(object, name)
-    object.fetch(name)
-  end
-  alias fetch_int fetch_entry
-  alias fetch_string fetch_entry
-end
 
 describe 'publishing messages (JSON)', functional: true do
   include_context 'JSON format'
